pyproject.toml
[tool.poetry]
name = "postgres-da-ai-agent"
version = "0.1.0"
description = "This project is a fork from the multi-agent-system from indydevdan, we expanded it's capabilities to connect with PrestoDB clusters."
authors = ["apereiracv <alejandro@axeom.xyz>", "alvropena <alvaro@axeom.xyz>"]
readme = "README.md"
packages = [{include = "postgres_da_ai_agent"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
openai = "^1.2.3"
psycopg2-binary = "^2.9.8"
argparse = "^1.4.0"
python-dotenv = "^1.0.0"
transformers = "^4.34.1"
torch = "^2.1.0"
scikit-learn = "^1.3.1"
tiktoken = "^0.5.1"
guidance = "^0.0.64"
presto-python-client = "^0.8.4"
pyautogen = "^0.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# Not available in current version due to pyautogen not supporting openai ^1.2.3
# old_start = "da_ai_agent.main:main"
start_presto = "postgres_da_ai_agent.main_presto:main"
start_postgres = "postgres_da_ai_agent.main_postgres:main"
turbo_presto = "postgres_da_ai_agent.turbo_main_presto:main"
turbo_postgres = "postgres_da_ai_agent.turbo_main_postgres:main"
